roll load-actual up into build-ideal-tree
